+++ doc/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 hachoir 3.0a2
 =============
 
+* metadata: get comment from ZIP
 * Fix ResourceWarning warnings on files. Add a close() method and support for
   the context manager protocol ("with obj: ...") to parsers, input and output
   streams.